Update to latest vertx-pac4j with vert.x 3.3.3 and pac4j 1.9.2
An API compatibility break between pac4j 1.9.1 and 1.9.2 meant that it was necessary to make a small change to the CAS client configuration code to enable compilation.

In addition, the SAML behaviour has changed so it attempts to write out a keystore file if it doesn't find one. Previously we were loading the file from the classpath (which meant that when built it was in the jar, but this isn't writeable, and for some reason the file was not being found, leading to a problem on write). Therefore we've now moved the keystore out into the external filesystem (as is more appropriate, it wouldn't be normal to build the keystore into the application).