Issue #38 Set content type header appropriately.
Vertx-pac4j-demo wasn't setting the content type header on most of its responses, leading to incorrect rendering when testing using Safari.

This commit will, if applied, set the content type response header to text/html where we are returning html, and to application/json when we return json to the caller.