Issue #39 Update profiles display and update wording
Update the section of the index and protected index which used to display the profile, and modify it to show a list of all authenticated profiles, as part of multiprofile support.

Update the wording around the custom authorizer facebook index to make it clear how to use.

This commit does not in any way complete issue #39, it is just a step on the route.